requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
email-validator==2.2.0
aiosmtplib==4.0.0
python-dotenv==1.0.1
starlette==0.46.1
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
gunicorn==23.0.0
backoff==2.2.1
pyaml-env==1.2.2
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-jose[cryptography]==3.4.0
alembic==1.15.1